2/3 split fastpath code for makereq into separate file (#204)
* split fastpath code for makereq into separate file
