+++ pxd.c
@@ -382,6 +382,19 @@ static const struct block_device_operations pxd_bd_ops = {
 	.release		= pxd_release,
 };
 
+#if defined(__PXD_BIO_MAKEREQ__) && LINUX_VERSION_CODE >= KERNEL_VERSION(5,9,0)
+static const struct block_device_operations pxd_bd_fpops = {
+    .owner          = THIS_MODULE,
+    .open           = pxd_open,
+    .release        = pxd_release,
+    .submit_bio     = pxd_bio_make_request_entryfn,
+};
+#define get_bd_fpops() (&pxd_bd_fpops)
+#else
+#define get_bd_fpops() (&pxd_bd_ops)
+#endif
+
+
 static void pxd_update_stats(struct fuse_req *req, int rw, unsigned int count)
 {
 		struct pxd_device *pxd_dev = req->queue->queuedata;
@@ -412,6 +425,7 @@ static bool __pxd_device_qfull(struct pxd_device *pxd_dev)
 	if (atomic_cmpxchg(&pxd_dev->congested, 1, 0) == 1) {
 		pxd_dev->nr_congestion_off++;
 	}
+
 	return 0;
 }
 
@@ -444,28 +458,28 @@ void pxd_check_q_decongested(struct pxd_device *pxd_dev)
 	}
 }
 
-static void pxd_request_complete(struct fuse_conn *fc, struct fuse_req *req)
+static void pxd_request_complete(struct fuse_conn *fc, struct fuse_req *req, int status)
 {
 	atomic_dec(&req->pxd_dev->ncount);
 	pxd_check_q_decongested(req->pxd_dev);
-	pxd_printk("%s: receive reply to %px(%lld) at %lld\n",
+	pxd_printk("%s: receive reply to %px(%lld) at %lld err %d\n",
 			__func__, req, req->in.unique,
-			req->pxd_rdwr_in.offset);
+			req->pxd_rdwr_in.offset, status);
 }
 
 static bool pxd_process_read_reply(struct fuse_conn *fc, struct fuse_req *req,
-	int status)
+		int status)
 {
 	trace_pxd_reply(req->in.unique, 0u);
 	pxd_update_stats(req, 0, BIO_SIZE(req->bio) / SECTOR_SIZE);
 	BIO_ENDIO(req->bio, status);
-	pxd_request_complete(fc, req);
+	pxd_request_complete(fc, req, status);
 
 	return true;
 }
 
 static bool pxd_process_write_reply(struct fuse_conn *fc, struct fuse_req *req,
-	int status)
+		int status)
 {
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(4,8,0) || defined(REQ_PREFLUSH)
 	trace_pxd_reply(req->in.unique, REQ_OP_WRITE);
@@ -474,16 +488,16 @@ static bool pxd_process_write_reply(struct fuse_conn *fc, struct fuse_req *req,
 #endif
 	pxd_update_stats(req, 1, BIO_SIZE(req->bio) / SECTOR_SIZE);
 	BIO_ENDIO(req->bio, status);
-	pxd_request_complete(fc, req);
+	pxd_request_complete(fc, req, status);
 
 	return true;
 }
 
 /* only used by the USE_REQUESTQ_MODEL definition */
 static bool pxd_process_read_reply_q(struct fuse_conn *fc, struct fuse_req *req,
-	int status)
+		int status)
 {
-	pxd_request_complete(fc, req);
+	pxd_request_complete(fc, req, status);
 #ifndef __PX_BLKMQ__
 	blk_end_request(req->rq, status, blk_rq_bytes(req->rq));
 	return true;
@@ -495,9 +509,10 @@ static bool pxd_process_read_reply_q(struct fuse_conn *fc, struct fuse_req *req,
 
 /* only used by the USE_REQUESTQ_MODEL definition */
 static bool pxd_process_write_reply_q(struct fuse_conn *fc, struct fuse_req *req,
-	int status)
+		int status)
 {
-	pxd_request_complete(fc, req);
+
+	pxd_request_complete(fc, req, status);
 #ifndef __PX_BLKMQ__
 	blk_end_request(req->rq, status, blk_rq_bytes(req->rq));
 	return true;
@@ -563,7 +578,7 @@ int pxd_handle_device_limits(struct fuse_req *req, uint32_t *size, uint64_t *off
 	struct request_queue *q = req->pxd_dev->disk->queue;
 	sector_t max_sectors, rq_sectors;
 
-	if (req->pxd_dev->using_blkque) {
+	if (!fastpath_enabled(req->pxd_dev)) {
 		return 0;
 	}
 
@@ -623,7 +638,7 @@ static int pxd_read_request(struct fuse_req *req, uint32_t size, uint64_t off,
 	}
 
 	req->in.opcode = PXD_READ;
-	if (!req->pxd_dev->using_blkque) {
+	if (fastpath_enabled(req->pxd_dev)) {
 		req->end = pxd_process_read_reply;
 	} else {
 		req->end = pxd_process_read_reply_q;
@@ -648,7 +663,7 @@ static int pxd_write_request(struct fuse_req *req, uint32_t size, uint64_t off,
 	}
 
 	req->in.opcode = PXD_WRITE;
-	if (!req->pxd_dev->using_blkque) {
+	if (fastpath_enabled(req->pxd_dev)) {
 		req->end = pxd_process_write_reply;
 	} else {
 		req->end = pxd_process_write_reply_q;
@@ -677,7 +692,7 @@ static int pxd_discard_request(struct fuse_req *req, uint32_t size, uint64_t off
 	}
 
 	req->in.opcode = PXD_DISCARD;
-	if (!req->pxd_dev->using_blkque) {
+	if (fastpath_enabled(req->pxd_dev)) {
 		req->end = pxd_process_write_reply;
 	} else {
 		req->end = pxd_process_write_reply_q;
@@ -702,7 +717,7 @@ static int pxd_write_same_request(struct fuse_req *req, uint32_t size, uint64_t 
 	}
 
 	req->in.opcode = PXD_WRITE_SAME;
-	if (!req->pxd_dev->using_blkque) {
+	if (fastpath_enabled(req->pxd_dev)) {
 		req->end = pxd_process_write_reply;
 	} else {
 		req->end = pxd_process_write_reply_q;
@@ -823,7 +838,7 @@ int pxd_initiate_ioswitch(struct pxd_device *pxd_dev, int code)
 {
 	struct fuse_req *req;
 
-	if (pxd_dev->using_blkque) {
+	if (!fastpath_enabled(pxd_dev)) {
 		return -EINVAL;
 	}
 
@@ -838,11 +853,7 @@ int pxd_initiate_ioswitch(struct pxd_device *pxd_dev, int code)
 
 	req->in.opcode = code;
 	req->end = pxd_process_ioswitch_complete;
-
-	req->pxd_rdwr_in.dev_minor = pxd_dev->minor;
-	req->pxd_rdwr_in.offset = 0;
-	req->pxd_rdwr_in.size = 0;
-	req->pxd_rdwr_in.flags = PXD_FLAGS_SYNC;
+	pxd_req_misc(req, 0, 0, pxd_dev->minor, PXD_FLAGS_SYNC);
 
 	fuse_request_send_nowait(&pxd_dev->ctx->fc, req);
 	return 0;
@@ -852,7 +863,7 @@ int pxd_initiate_failover(struct pxd_device *pxd_dev)
 {
 	int rc;
 
-	if (!pxd_dev->fp.fastpath) {
+	if (!fastpath_active(pxd_dev)) {
 		// already in native path
 		return -EINVAL;
 	}
@@ -880,7 +891,7 @@ int pxd_initiate_fallback(struct pxd_device *pxd_dev)
 {
 	int rc;
 
-	if (pxd_dev->fp.fastpath) {
+	if (fastpath_active(pxd_dev)) {
 		// already in fast path
 		return -EINVAL;
 	}
@@ -933,50 +944,6 @@ void pxd_reroute_slowpath(struct request_queue *q, struct bio *bio)
 	}
 
 	fuse_request_send_nowait(&pxd_dev->ctx->fc, req);
-}
-
-// fastpath uses this path to punt requests to slowpath
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(4,4,0)
-blk_qc_t pxd_make_request_slowpath(struct request_queue *q, struct bio *bio)
-#define BLK_QC_RETVAL BLK_QC_T_NONE
-#else
-void pxd_make_request_slowpath(struct request_queue *q, struct bio *bio)
-#define BLK_QC_RETVAL
-#endif
-{
-	struct pxd_device *pxd_dev = q->queuedata;
-	struct fuse_req *req;
-
-	pxd_printk("%s: dev m %d g %lld %s at %ld len %d bytes %d pages "
-			"flags 0x%x op_flags 0x%x\n", __func__,
-			pxd_dev->minor, pxd_dev->dev_id,
-			bio_data_dir(bio) == WRITE ? "wr" : "rd",
-			BIO_SECTOR(bio) * SECTOR_SIZE, BIO_SIZE(bio),
-			bio->bi_vcnt, bio->bi_flags, get_op_flags(bio));
-
-	req = pxd_fuse_req(pxd_dev);
-	if (IS_ERR_OR_NULL(req)) {
-		bio_io_error(bio);
-		return BLK_QC_RETVAL;
-	}
-
-	req->pxd_dev = pxd_dev;
-	req->bio = bio;
-	req->queue = q;
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(4,8,0) || defined(REQ_PREFLUSH)
-	if (pxd_request(req, BIO_SIZE(bio), BIO_SECTOR(bio) * SECTOR_SIZE,
-		pxd_dev->minor, bio_op(bio), bio->bi_opf)) {
-#else
-	if (pxd_request(req, BIO_SIZE(bio), BIO_SECTOR(bio) * SECTOR_SIZE,
-		    pxd_dev->minor, bio->bi_rw)) {
-#endif
-		fuse_request_free(req);
-		bio_io_error(bio);
-		return BLK_QC_RETVAL;
-	}
-
-	fuse_request_send_nowait(&pxd_dev->ctx->fc, req);
-	return BLK_QC_RETVAL;
 }
 
 #if !defined(__PX_BLKMQ__)
@@ -1099,19 +1066,20 @@ static int pxd_init_disk(struct pxd_device *pxd_dev, struct pxd_add_ext_out *add
 
 	// only fastpath uses direct block io process bypassing request queuing
 #ifndef __PX_FASTPATH__
-	if (!pxd_dev->using_blkque) {
+	if (fastpath_enabled(pxd_dev)) {
 		printk(KERN_NOTICE"PX driver does not support fastpath, disabling it\n");
-		pxd_dev->using_blkque = true;
+		pxd_dev->fastpath = false;
 	}
-#else
-	if (!pxd_dev->using_blkque) {
+#endif
+	if (fastpath_enabled(pxd_dev)) {
 		pxd_printk("adding disk for fastpath device %llu", pxd_dev->dev_id);
+		disk->fops = get_bd_fpops();
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(5,9,0)
 	  q = blk_alloc_queue(NUMA_NO_NODE);
 #elif LINUX_VERSION_CODE >= KERNEL_VERSION(5,7,0)
-		q = blk_alloc_queue(pxd_make_request_fastpath, NUMA_NO_NODE);
+		q = blk_alloc_queue(pxd_bio_make_request_entryfn, NUMA_NO_NODE);
 #elif LINUX_VERSION_CODE == KERNEL_VERSION(4,18,0) && defined(__EL8__) && defined(QUEUE_FLAG_NOWAIT)
-        q = blk_alloc_queue_rh(pxd_make_request_fastpath, NUMA_NO_NODE);
+        q = blk_alloc_queue_rh(pxd_bio_make_request_entryfn, NUMA_NO_NODE);
 #else
 		q = blk_alloc_queue(GFP_KERNEL);
 #endif
@@ -1125,10 +1093,9 @@ static int pxd_init_disk(struct pxd_device *pxd_dev, struct pxd_add_ext_out *add
 #endif
 
 #if LINUX_VERSION_CODE < KERNEL_VERSION(5,7,0) && !defined(QUEUE_FLAG_NOWAIT)
-		blk_queue_make_request(q, pxd_make_request_fastpath);
+		blk_queue_make_request(q, pxd_bio_make_request_entryfn);
 #endif
 	} else {
-#endif
 
 #ifdef __PX_BLKMQ__
 	  memset(&pxd_dev->tag_set, 0, sizeof(pxd_dev->tag_set));
@@ -1156,10 +1123,8 @@ static int pxd_init_disk(struct pxd_device *pxd_dev, struct pxd_add_ext_out *add
 	  	goto out_disk;
 	  }
 #endif
-
-#ifdef __PX_FASTPATH__
 	}
-#endif
+
 	blk_queue_max_hw_sectors(q, SEGMENT_SIZE / SECTOR_SIZE);
 	blk_queue_max_segment_size(q, SEGMENT_SIZE);
 	blk_queue_max_segments(q, (SEGMENT_SIZE / PXD_LBS));
@@ -1187,7 +1152,7 @@ static int pxd_init_disk(struct pxd_device *pxd_dev, struct pxd_add_ext_out *add
 	BLK_QUEUE_FLUSH(q);
 
 	/* adjust queue limits to be compatible with backing device */
-	if (!pxd_dev->using_blkque) {
+	if (fastpath_enabled(pxd_dev)) {
 		pxd_fastpath_adjust_limits(pxd_dev, q);
 	}
 
@@ -1208,13 +1173,14 @@ static void pxd_free_disk(struct pxd_device *pxd_dev)
 	if (!disk)
 		return;
 
+	pxd_fastpath_cleanup(pxd_dev);
 	pxd_dev->disk = NULL;
 	if (disk->flags & GENHD_FL_UP) {
 		del_gendisk(disk);
 		if (disk->queue)
 			blk_cleanup_queue(disk->queue);
 #ifdef __PX_BLKMQ__
-		if (pxd_dev->using_blkque) blk_mq_free_tag_set(&pxd_dev->tag_set);
+		if (!fastpath_enabled(pxd_dev)) blk_mq_free_tag_set(&pxd_dev->tag_set);
 #endif
 	}
 	put_disk(disk);
@@ -1294,7 +1260,7 @@ ssize_t pxd_add(struct fuse_conn *fc, struct pxd_add_ext_out *add)
 	pxd_dev->connected = true; // fuse slow path connection
 	pxd_dev->size = add->size;
 	pxd_dev->mode = add->open_mode;
-	pxd_dev->using_blkque = !add->enable_fp;
+	pxd_dev->fastpath = add->enable_fp;
 
 	// congestion init
 	init_waitqueue_head(&pxd_dev->suspend_wq);
@@ -1313,7 +1279,7 @@ ssize_t pxd_add(struct fuse_conn *fc, struct pxd_add_ext_out *add)
 	if (err)
 		goto out_id;
 
-	if (add->enable_fp) {
+	if (fastpath_enabled(pxd_dev)) {
 		err = pxd_init_fastpath_target(pxd_dev, &add->paths);
 		if (err) {
 			pxd_fastpath_cleanup(pxd_dev);
@@ -1323,6 +1289,7 @@ ssize_t pxd_add(struct fuse_conn *fc, struct pxd_add_ext_out *add)
 
 	err = pxd_init_disk(pxd_dev, add);
 	if (err) {
+		pxd_fastpath_cleanup(pxd_dev);
 		goto out_id;
 	}
 
@@ -1486,14 +1453,14 @@ ssize_t pxd_read_init(struct fuse_conn *fc, struct iov_iter *iter)
 	copied += sizeof(pxd_init);
 
 	list_for_each_entry(pxd_dev, &ctx->list, node) {
-		struct pxd_dev_id id = {0};
+		struct pxd_dev_id id = {0}; // initialize all fields to zero
 		id.dev_id = pxd_dev->dev_id;
 		id.local_minor = pxd_dev->minor;
 		id.fastpath = 0;
-		id.blkmq_device = 0;
 		id.suspend = 0;
 		if (pxd_dev->fp.fastpath) id.fastpath = 1;
-		if (pxd_dev->using_blkque) id.blkmq_device = 1;
+		id.blkmq_device = 1;
+		if (fastpath_enabled(pxd_dev)) id.blkmq_device = 0;
 		if (atomic_read(&pxd_dev->fp.app_suspend)) id.suspend = 1;
 		if (copy_to_iter(&id, sizeof(id), iter) != sizeof(id)) {
 			printk(KERN_ERR "%s: copy dev id error copied %ld\n", __func__,
@@ -1518,10 +1485,10 @@ copy_error:
 
 static int __pxd_update_path(struct pxd_device *pxd_dev, struct pxd_update_path_out *update_path)
 {
-	if (pxd_dev->using_blkque) {
+	if (!fastpath_enabled(pxd_dev)) {
 		printk(KERN_WARNING"%llu: block device registered for native path - cannot update for fastpath\n", pxd_dev->dev_id);
 		return -EINVAL;
-	} else if (pxd_dev->fp.fastpath) {
+	} else if (fastpath_active(pxd_dev)) {
 		printk(KERN_ERR"%llu: device already in fast path - cannot update\n", pxd_dev->dev_id);
 		return -EINVAL;
 	}
@@ -1807,8 +1774,8 @@ static ssize_t pxd_debug_show(struct device *dev,
 	int suspend;
 
 	suspend=pxd_suspend_state(pxd_dev);
-	return sprintf(buf, "nfd:%d,suspend:%d,fastpath:%d,mqdevice:%d,app_suspend:%d\n",
-			pxd_dev->fp.nfd, suspend, pxd_dev->fp.fastpath, pxd_dev->using_blkque,
+	return sprintf(buf, "nfd:%d,suspend:%d,fpenabled:%d,fpactive:%d,app_suspend:%d\n",
+			pxd_dev->fp.nfd, suspend, fastpath_enabled(pxd_dev), fastpath_active(pxd_dev),
 			atomic_read(&pxd_dev->fp.app_suspend));
 }
