+++ pxd_compat.h
@@ -46,13 +46,17 @@
 #else
 #define BIO_OP(bio)   ((bio)->bi_rw)
 // only supports read or write
-#define SUBMIT_BIO(bio)  submit_bio(((bio)->bi_rw & 1), bio)
+#define BIO_OP(bio)   ((bio)->bi_rw & 1)
+#define SUBMIT_BIO(bio)  submit_bio(BIO_OP(bio), bio)
 #endif
 
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(4,13,0)
-#define BIOSET_CREATE(sz, pad)   bioset_create(sz, pad, 0)
+#define BIOSET_CREATE(sz, pad, opt)   bioset_create(sz, pad, opt)
 #else
-#define BIOSET_CREATE(sz, pad)   bioset_create(sz, pad)
+#ifndef BIOSET_NEED_BVECS
+#define BIOSET_NEED_BVECS (1)
+#endif
+#define BIOSET_CREATE(sz, pad, opt)   bioset_create(sz, pad)
 #endif
 
 #if defined(bio_set_dev)
@@ -121,7 +125,7 @@ static inline unsigned int get_op_flags(struct bio *bio)
 #elif LINUX_VERSION_CODE < KERNEL_VERSION(4,10,0)
 	op_flags = bio_flags(bio);
 #else
-	op_flags = ((bio->bi_opf & ~REQ_OP_MASK) >> REQ_OP_BITS);
+	op_flags = (bio->bi_opf & ~REQ_OP_MASK);
 #endif
 	return op_flags;
 }
