+++ pxd_core.h
@@ -48,7 +48,7 @@ struct pxd_device {
 	struct pxd_context *ctx;
 	bool connected;
 	mode_t mode;
-	bool using_blkque; // this is persistent, how the block device registered with kernel
+	bool fastpath; // this is persistent, how the block device registered with kernel
 
 #define PXD_ACTIVE(pxd_dev)  (atomic_read(&pxd_dev->ncount))
 	// congestion handling
@@ -63,6 +63,18 @@ struct pxd_device {
         struct blk_mq_tag_set tag_set;
 #endif
 };
+
+// how pxd_device got registered with the kernel during device add.
+static inline
+bool fastpath_enabled(struct pxd_device *pxd_dev) {
+	return pxd_dev->fastpath;
+}
+
+// current IO status - fastpath vs nativepath
+static inline
+bool fastpath_active(struct pxd_device *pxd_dev) {
+	return pxd_dev->fp.fastpath;
+}
 
 void pxd_check_q_congested(struct pxd_device *pxd_dev);
 void pxd_check_q_decongested(struct pxd_device *pxd_dev);
@@ -85,16 +97,11 @@ void pxd_check_q_decongested(struct pxd_device *pxd_dev);
 
 #define SEGMENT_SIZE (1024 * 1024)
 
-// slow path make request io entry point
-struct request_queue;
-struct bio;
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(4,4,0)
-blk_qc_t pxd_make_request_slowpath(struct request_queue *q, struct bio *bio);
-#else
-void pxd_make_request_slowpath(struct request_queue *q, struct bio *bio);
-#endif
-
+#ifdef __PXD_BIO_MAKEREQ__
 void pxd_reroute_slowpath(struct request_queue *q, struct bio *bio);
+#else
+void pxdmq_reroute_slowpath(struct fuse_req*);
+#endif
 int pxd_initiate_fallback(struct pxd_device *pxd_dev);
 int pxd_initiate_failover(struct pxd_device *pxd_dev);
