+++ pxd_fastpath.h
@@ -13,30 +13,11 @@
 #include <linux/falloc.h>
 #include <linux/bio.h>
 
+#include "pxd_bio.h"
+
 struct pxd_device;
 struct pxd_context;
 struct fuse_conn;
-
-// Added metadata for each bio
-struct pxd_io_tracker {
-#define PXD_IOT_MAGIC (0xbeefcafe)
-	unsigned int magic;
-	struct pxd_device *pxd_dev; // back pointer to pxd device
-	struct pxd_io_tracker *head; // back pointer to head copy [ALL]
-	struct list_head replicas; // only replica needs this
-	struct list_head item; // only HEAD needs this
-	atomic_t active; // only HEAD has refs to all active IO
-	struct file* file;
-
-	unsigned long start; // start time [HEAD]
-	struct bio *orig;    // original request bio [HEAD]
-	int status; // should be zero, non-zero indicates consolidated fail status
-
-	struct work_struct wi; // work item
-
-	// THIS SHOULD BE LAST ITEM
-	struct bio clone;    // cloned bio [ALL]
-};
 
 struct pxd_sync_ws {
 	struct work_struct ws;
@@ -93,17 +74,6 @@ void pxd_fastpath_cleanup(struct pxd_device *pxd_dev);
 
 void pxdctx_set_connected(struct pxd_context *ctx, bool enable);
 
-// IO entry point
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(5,9,0)
-blk_qc_t pxd_make_request_fastpath(struct bio *bio);
-#elif LINUX_VERSION_CODE >= KERNEL_VERSION(4,4,0)
-blk_qc_t pxd_make_request_fastpath(struct request_queue *q, struct bio *bio);
-#define BLK_QC_RETVAL BLK_QC_T_NONE
-#else
-void pxd_make_request_fastpath(struct request_queue *q, struct bio *bio);
-#define BLK_QC_RETVAL
-#endif
-
 void enableFastPath(struct pxd_device *pxd_dev, bool force);
 void disableFastPath(struct pxd_device *pxd_dev, bool skipSync);
 
@@ -130,8 +100,6 @@ int pxd_request_ioswitch(struct pxd_device *pxd_dev, int code);
 
 // handle IO reroutes and switch events
 void pxd_reissuefailQ(struct pxd_device *pxd_dev, struct list_head *ios, int status);
-void pxd_abortfailQ(struct pxd_device *pxd_dev);
-void __pxd_abortfailQ(struct pxd_device *pxd_dev);
 
 static inline
 struct block_device* get_bdev(struct file *fileh)
@@ -154,6 +122,20 @@ unsigned get_mode(struct file *fileh)
     struct inode *inode = mapping->host;
 
     return inode->i_mode;
+}
+
+static inline
+int remap_io_status(int status)
+{
+	switch (status) {
+	case 0: // success
+	case -EOPNOTSUPP: // op not supported - no failover
+	case -ENOSPC: // no space on device - no failover
+	case -ENOMEM: // no memory - no failover
+		return status;
+	}
+
+	return -EIO;
 }
 
 #endif /* _PXD_FASTPATH_H_ */
